send_message now uses size_t for len parameter.

+++ peer.c
@@ -180,7 +180,7 @@ int send_buffer(struct peer *p)
 	return 0;
 }
 
-int send_message(struct peer *p, char *rendered, uint32_t len)
+int send_message(struct peer *p, char *rendered, size_t len)
 {
 	int ret;
 	ssize_t sent;
